Array updation based on Next Greater Number
Given an array having distinct integer elements, the task is to find the first next greater element on right side for each element of the array. If no such element exists, output -1.
BUT, 
Update the given array in such a way that the indices where the present element if is smaller than element in right, should contain -1 and greater number if it exists on right.