Create MajorityElement.cpp
Given an array of size n, find the majority element. The majority element is the element that appears more than n/2 times.

You may assume that the array is non-empty and and the contents of array are always positive

Complete the function given getMajorityElement() in the editor, that accepts an array & returns the majority element of the array.