Create LLReverse_Itetrative.cpp
A linked list is a collection of nodes. First node is called head node and last node called the tail node.

Now, given a pointer to the head node of a linked list, can you reverse the linked list i.e. New list must start with tail node and end at head node. So it becomes the reverse of the original list.

Complete the function reverseList() which takes the head node of a linked list as a parameter and return the new head of reversed list.